Add wrap_content support
Modifies the height computation when the wrap_content is used. The height of the text, together to the vertical margin that can be set, is also included in the heightMeasureSpec that is passed. This will ensure the layout will have the correct height and can properly layout and draw the content